reworked commands for bot
