+++ lib/rotoworld.js
@@ -5,6 +5,7 @@ const querystring = require('querystring');
 // There are some players where there are multiple NFL players.
 // We're jerks, but we only care about the famous ones. This is an
 // exceptions list so we make sure we get the one that's important.
+
 const playerExceptions = [
     { name: 'Johnson, David', url: 'http://www.rotoworld.com/player/nfl/10404/david-johnson' },
     { name: 'Marshall, Brandon', url: 'http://www.rotoworld.com/player/nfl/3653/brandon-marshall' },
@@ -19,10 +20,10 @@ exports.getPlayer = function (playerName) {
             let [firstName, ...lastName] = playerName.toString().split(" ").slice(0);
             playerName = lastName + ', ' + firstName;
         }
-
         // Let's see if we have a player in the player exception list, otherwise we can just search as normal.
         getPlayerExceptions = playerExceptions.filter(function (value) {
-            return value.name == playerName;
+            // *tc216997* making sure it isn't case sensitive for player exceptions
+            return value.name.toLowerCase() == playerName.toLowerCase();
         });
         if (getPlayerExceptions.length == 1) {
             url = getPlayerExceptions[0].url;
