+++ lib/utilities.js
@@ -22,7 +22,7 @@ const eightBallAnswers = [
 ];
 
 utilities.chooseOne = function (chooseList) {
-    return chooseList.split(',')[Math.floor(Math.random() * chooseListArray.length)];
+    return chooseList[Math.floor(Math.random() * chooseList.length)];
 };
 
 utilities.eightBall = function () {
